Reverted animation changes, that approach needs re-thinking
