+++ src/components/Text/Text.tsx
@@ -5,7 +5,7 @@ import {
   TextStyle,
   Animated,
 } from "react-native";
-import { Animation, Flex, Position, Size, Spacing } from "../../models";
+import { Flex, Position, Size, Spacing } from "../../models";
 import {
   extractFlexProperties,
   extractPositionProperties,
@@ -26,7 +26,7 @@ interface OwnProps extends RNTextProps {
   opacity?: TextStyle["opacity"];
 }
 
-export type TextProps = OwnProps & Spacing & Size & Flex & Position & Animation;
+export type TextProps = OwnProps & Spacing & Size & Flex & Position;
 
 /** Create a Font.d.ts type in your typescript types directory and define fonts as follows:
  * @example type Font = "Helvetica" | "Montserrat" ...  */
@@ -42,7 +42,6 @@ export const Text: React.FC<PropsWithChildren<TextProps>> = ({
   opacity,
   margins,
   paddings,
-  animations,
   ...props
 }) => {
   const { defaultFont, defaultFontSize, defaultTextColor } = useSkeletor();
@@ -84,7 +83,6 @@ export const Text: React.FC<PropsWithChildren<TextProps>> = ({
         flexProps,
         positionProps,
         style,
-        animations,
       ]),
     [
       color,
