+++ src/hooks/useAnimation.ts
@@ -1,16 +1,28 @@
-import {
-  Animation,
-  AnimationConfiguration,
-  AnimationDefinition,
-  AnimationProperties,
-} from "models";
 import { useMemo, useRef } from "react";
 import { Animated, ViewStyle } from "react-native";
 
-interface AnimationSet<Keys extends keyof Partial<ViewStyle>> {
+/** Any is a hotfix, requires more investigation */
+export type Animation<Keys extends keyof Partial<ViewStyle>> = Record<
+  Keys,
+  Animated.AnimatedInterpolation<string | number> | any
+>;
+export type AnimationDefinition<Keys extends keyof Partial<ViewStyle>> = Record<
+  Keys,
+  number[] | string[]
+>;
+
+export interface AnimationConfiguration {
+  /** In miliseconds */
+  duration: number;
+  /** Loop will disable native driver because it breaks the loop animation (at least it did last time I tested in 2020.) */
+  loop?: boolean;
+  useNativeDriver?: boolean;
+}
+
+export interface AnimationSet<Keys extends keyof Partial<ViewStyle>> {
   values: Animated.Value[];
   definitions: AnimationDefinition<Keys>;
-  animations: AnimationProperties<Keys>;
+  animations: Animation<Keys>;
   configuration: AnimationConfiguration;
 }
 
@@ -23,7 +35,7 @@ export function useAnimation<Keys extends keyof Partial<ViewStyle>>(
   const values = useRef(keys.map(() => new Animated.Value(0))).current;
 
   const animations = useMemo(() => {
-    const result: Partial<AnimationProperties<Keys>> = {};
+    const result: Partial<Animation<Keys>> = {};
 
     keys.forEach((key, index) => {
       const value = values[index];
@@ -38,7 +50,7 @@ export function useAnimation<Keys extends keyof Partial<ViewStyle>>(
       result[key] = interpolation;
     });
 
-    return result as AnimationProperties<Keys>;
+    return result as Animation<Keys>;
   }, [styles]);
 
   return {
